cxr-platform/orchestrator/requirements.txt
fastapi>=0.95.0,<0.110
uvicorn[standard]>=0.21.0,<0.29
httpx>=0.23.0,<0.28
boto3>=1.26.0,<1.35
pymysql>=1.0.2,<1.2
python-dotenv>=1.0.0,<1.1
pydantic>=2.0.0,<3.0.0
pydantic-settings>=2.0.0,<3.0.0
pydicom>=2.3.0,<3.0.0
Pillow>=9.0.0,<11.0.0
numpy>=1.21.0,<2.0.0